% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty_greenland.R
\docType{data}
\name{dirty_greenland}
\alias{dirty_greenland}
\title{Dirty_Greenland dataset}
\format{A data frame containing 263 columns and 16 variables.}
\description{
Water nutrient dataset from the Redfish 2013 cruise that took place in the Irminger Sea.
All measurements were made and reported by Ryan Barkhouse of Dalhousie University. This is the dirty version of "greenland"
because it contains error values.
}
\references{
Redfish 2013, obtained from: \url{https://www.dal.ca/diff/cerc.html}
}
\keyword{data}
