% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenland.R
\docType{data}
\name{greenland}
\alias{greenland}
\title{Greenland dataset}
\format{A data frame containing 263 columns and 16 variables.}
\description{
Water nutrient dataset from the Redfish 2013 cruise that took place in the Irminger Sea.
All measurements were made and reported by Ryan Barkhouse of Dalhousie University.
}
\references{
Redfish 2013, obtained from: \url{https://www.dal.ca/diff/cerc.html}
}
\keyword{data}
