% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrientRange.R
\name{nutrientRange}
\alias{nutrientRange}
\title{Geneate nutrient range stats}
\usage{
nutrientRange(depth_column, nutrient_columns, depth1, depth2, depth3, depth4)
}
\arguments{
\item{depth_column}{The depth (or pressure) values.}

\item{nutrient_columns}{All nutrient data. Must be provided in separate columns.}

\item{depth1}{The upper limit of your shallow depth range.}

\item{depth2}{The lower limit of your shallow depth range.}

\item{depth3}{The upper limit of your deep depth range.}

\item{depth4}{The lower limit of your deep depth range.}
}
\description{
This function produces nutrient ranges at two different depths - usually at the surface and at depth. The depths are given to the function as two
separate depth ranges. The result of \code{nutrientRange} is stored under \code{nutrientRanges} in the global environment.
}
\examples{

library(oceanchemistry)

data("greenland", package = "oceanchemistry")
head(greenland)

nutrientRange(greenland$Depth,
     greenland[,10:11],
     0,15,80,115)

View(nutrientRanges)
}
\keyword{Nutrient,}
\keyword{column}
\keyword{range,}
\keyword{water}
