% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanchemistry-package.r
\docType{package}
\name{oceanchemistry}
\alias{oceanchemistry}
\alias{oceanchemistry-package}
\title{Ocean chemical analysis for R}
\description{
The oceanchemistry package helps ocean chemists who are working with measurements from the water column. The functions in this package are
limited to chemical oceanography.
}
\section{Audience}{


This package is built for ocean chemists. Most of the functions in this package are meant for those who do not want to write their own code.
As a result, this package's functions do not offer users much freedom.
}

\section{Motivation}{

The idea for the oceanchemistry package came from a project that had some very messy data. Oceandata View (ODV), the industry standard for analyzing ocean chemical
data, could not process the data. This is probably because the software was not familiar with the layout of the data. Instead of
reformatting the data in Excel, this stubborn summer student decided to do everything in R.
}

