% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redfield.R
\name{redfield}
\alias{redfield}
\title{Redfield ratio analysis}
\usage{
redfield(C, N, P, subset = x)
}
\arguments{
\item{C}{Drawdown of organic carbon in umol/kg.}

\item{N}{Drawdown of nitrate in umol/kg.}

\item{P}{Drawdown of phosphate in umol/kg.}

\item{subset}{Subset the data by any variables i.e. cruise, station.}
}
\description{
This function provides redfield ratios and statistics. Any \code{NA} values are automatically removed. The function accepts the optional argument
\code{subset}, which lets you specify whether the redfield statistics should be organized by any particular variable.
}
\examples{

library(oceanchemistry)

data("greenland", package = "oceanchemistry")
head(greenland)

redfield(greenland$'Carbon drawdown',
     greenland$'Nitrate drawdown',
     greenland$'Phosphate drawdown',
     subset = greenland$Cruise)

View(redfield1)
View(redstats1)
}
\keyword{Redfield,}
\keyword{ocean}
