% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redplot.R
\name{redplot}
\alias{redplot}
\title{Generate redfield plots}
\usage{
redplot(C, N, P, subset = x, lwd = 2, size = 3, font.size = 23,
  lty = "solid")
}
\arguments{
\item{C}{Drawdown of organic carbon in umol/kg.}

\item{N}{Drawdown of nitrate in umol/kg.}

\item{P}{Drawdown of phosphate in umol/kg.}

\item{subset}{Subset the data by any variables i.e. cruise, station.}

\item{lwd}{Optional graphic parameter.}

\item{size}{Optional graphic parameter.}

\item{font.size}{Optional graphic parameter.}

\item{lty}{Optional graphic parameter.}
}
\description{
This function generates redfield plots.These plots are a visual comparison between the drawdown of major nutrients to the corresponding
terms of the redfield ratio of C:N:P. This function produces three scatter plots depitcting (1) C:N, (2) C:P, (3) N:P. All three
plots include a reference line that represents the redfield ratio.\cr\cr
The results of this function are printed for you. Each plot is also stored in the global environment if you wish to adjust its
presentation or parameters. \cr \cr
You may subset the redfield data if you want to differentiate between the points on the scatter plot. This is usually used
to separate measurements from different stations or cruises.
}
\examples{

library(oceanchemistry)

data("greenland", package = "oceanchemistry")
head(greenland)

redplot(greenland$'Carbon drawdown',
     greenland$'Nitrate drawdown',
     greenland$'Phosphate drawdown',
     subset = greenland$Cruise)

}
\seealso{
This function relies on ggplot2 (\url{http://ggplot2.tidyverse.org/reference/}) for plotting and gridExtra (\url{https://cran.r-project.org/web/packages/gridExtra/gridExtra.pdf})
for arranging the plots. Please refer to these packages for any plot customization.
}
\keyword{Redfield,}
\keyword{ocean}
