% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubber.R
\name{scrubber}
\alias{scrubber}
\title{A data scrubber}
\usage{
scrubber(dataframe, replacee, replacer = NA)
}
\arguments{
\item{dataframe}{The dataset to be cleaned.}

\item{replacee}{The error values to be replace.}

\item{replacer}{The value used to replace your error values. This is automativally set to \code{NA}, but can be changed.}
}
\description{
This function helps you remove error values from your data. It is common for oceanographic instrumentation to report error values
during a project. Typically these are \code{-999} or \code{0}. Whatever your error value is, this function helps replace these values
with \code{NA}. This effectively removes error values from the dataset. The clean data is saved under \code{dataset} in the global environment.
}
\examples{
library(oceanchemistry)

data("dirty_greenland")
head(dirty_greenland)

scrubber(dirty_greenland, c(-999,0))


}
\keyword{Scrubber,}
\keyword{data,}
\keyword{dataset,}
\keyword{error}
\keyword{ocean,}
