% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfacePlot.R
\name{surfacePlot}
\alias{surfacePlot}
\title{Generate a contour or image}
\usage{
surfacePlot(depth_column, latitude, longitude, measurements, plot_depth,
  surface = T, contour = F)
}
\arguments{
\item{depth_column}{Depth (or pressure) values.}

\item{latitude}{Latitude coordinates.}

\item{longitude}{Longitude coordinates.}

\item{measurements}{The measurements to plot. If you have multiple variables they should be presented in columns or a dataframe.}

\item{plot_depth}{The lower limit of the surface plot. Currently this function will plot between the surface and whatever depth is input.}

\item{surface}{Do you want to make a surface plot?}

\item{contour}{Do you want to make a contour plot?}
}
\description{
This function generates a contour plot or an images (surface plot). The most frequent application of this function is surface nutrient
data. \code{surfacePlot} can handle multiple measurements or variables at once. You can choose between a surface plot,
a contour plot, or both. The surface plot uses a colour-blind-friendly palette taken from the \code{oce} package.
}
\examples{
library(oceanchemistry)

data("greenland", package = "oceanchemistry")
head(greenland)

surfacePlot(greenland$Depth,
     greenland$Latitude,
     greenland$Longitude,
     data.frame(greenland$`NO2 + NO3 [umol/lg]`),
     plot_depth = 10)


}
\seealso{
This function relies on the \code{interpBarnes} function from the \code{oce}
package: \url{https://cran.r-project.org/web/packages/oce/oce.pdf}.
}
\keyword{Contour,}
\keyword{measurements}
\keyword{ocean,}
\keyword{surface,}
